requirements.txt
asgiref==3.8.1
Django==5.0.4
django-phone-field==1.8.1
django-phonenumber-field==7.3.0
djangorestframework==3.15.1
phonenumbers==8.13.35
psycopg==3.1.18
sqlparse==0.5.0
typing_extensions==4.11.0
